refact: cleanup chain command code
