+++ packages/contract/src/__test__/contract-new.test.ts
@@ -2,8 +2,9 @@ import { NativeAssetId, ZeroBytes32 } from '@fuel-ts/constants';
 import { Provider } from '@fuel-ts/providers';
 import { TestUtils } from '@fuel-ts/wallet';
 
+import Contract from '../contract-new';
+
 import { setup } from './call-test-contract/call-test-contract.test';
-import Contract from './contract';
 
 const jsonFragment = {
   type: 'function',
