+++ packages/contract/src/__test__/token-test-contract/token_abi/src/main.sw
