+++ packages/contract/src/__test__/token-test-contract/token_contract/src/main.sw
